Revert "Rollback commons change until bugs are fixed (#654)"
This reverts commit 05617084128e564dd25675c205f191d4c8125457.

+++ EU4ToVic3/Source/V3World/EconomyManager/Demand/MarketTracker.cpp
@@ -156,7 +156,7 @@ void V3::MarketTracker::integrateBuilding(const Building& building,
 
 		lostSubsistence += marketJobs.createJobs(urbanResources.getJobs(),
 			 subsistenceResources.getJobs(),
-			 p * urbanFrac,
+			 static_cast<int>(p * urbanFrac),
 			 defaultRatio,
 			 addedWorkingPopFraction,
 			 {},
